// server/prisma/generated/zod/schemas/objects/ProductSumOrderByAggregateInput.schema.ts
import { z } from 'zod';
import { SortOrderSchema } from '../enums/SortOrder.schema';

import type { Prisma } from '@prisma/client';

const Schema: z.ZodType<Prisma.ProductSumOrderByAggregateInput> = z
  .object({
    id: z.lazy(() => SortOrderSchema).optional(),
    stock: z.lazy(() => SortOrderSchema).optional(),
    price: z.lazy(() => SortOrderSchema).optional(),
    discount: z.lazy(() => SortOrderSchema).optional(),
    peso: z.lazy(() => SortOrderSchema).optional(),
    anhoEdicion: z.lazy(() => SortOrderSchema).optional(),
    numeroPaginas: z.lazy(() => SortOrderSchema).optional(),
    sellerId: z.lazy(() => SortOrderSchema).optional(),
  })
  .strict();

export const ProductSumOrderByAggregateInputObjectSchema = Schema;